Initial Load of the AnsiWrap Script
This is a wrapper for the ANSP program that performs some filesystem
check before creating the ANSP file so that the previous ANSIP file is
not overwritten before submission and archive.  This wrapper script has
the option to run some post ANSIP file processing to certain things.
This part is optional.